+++ lib/src/features/home/pages/cache_chat_page/components/chat_content/elements/chat_image_item.dart
@@ -11,6 +11,7 @@ import 'package:studtime/src/features/home/pages/cache_chat_page/blocs/chat_cont
 import 'package:studtime/src/shared/data/models/chat/chat_image.dart';
 import 'package:studtime/src/shared/extensions/on_datetime.dart';
 import 'package:studtime/src/shared/styles/app_paddings.dart';
+import 'package:studtime/src/shared/widgets/dialogs/confirm_delete.dart';
 
 class ChatImageItem extends StatelessWidget {
   const ChatImageItem({
@@ -22,6 +23,8 @@ class ChatImageItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final chatContentCubit = context.read<ChatContentCubit>();
+
     return CupertinoButton(
       padding: const EdgeInsets.all(0.0),
       onPressed: () => OpenFilex.open(chatImage.path),
@@ -29,9 +32,11 @@ class ChatImageItem extends StatelessWidget {
         key: ObjectKey(chatImage),
         trailingActions: <SwipeAction>[
           SwipeAction(
-            onTap: (CompletionHandler handler) {
-              context
-                  .read<ChatContentCubit>()
+            onTap: (CompletionHandler handler) async {
+              final isConfirmed = await confirmDelete(context);
+              if (isConfirmed != true) return;
+
+              chatContentCubit
                   .deleteMessage(chatImage)
                   .then((value) => handler(true))
                   .onError(
